+++ src/arms_gibbs.cpp
@@ -62,7 +62,8 @@ void ARMS_Gibbs::arms_gibbs_xi(
     mydata->datProportion = datProportion.memptr();
     mydata->weibullS = weibullS.memptr();
     mydata->datX = dataclass.datX0.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
 
     // define how many ARMS samples to draw for currentPars
     // arma::mat samp = arma::zeros<arma::mat>(p, armsPar.n + 1);
@@ -237,7 +238,8 @@ void ARMS_Gibbs::arms_gibbs_beta(
     mydata->datMu = datMu.memptr();
     mydata->datProportion = datProportion.memptr();
     mydata->weibullS = weibullS.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
     mydata->datTime = dataclass.datTime.memptr();
 
     // arma::vec tauSq_tmp = arma::ones<arma::vec>(L);
@@ -412,7 +414,8 @@ void ARMS_Gibbs::arms_gibbs_betaK(
     mydata->datMu = datMu.memptr();
     mydata->datProportion = datProportion.memptr();
     mydata->weibullS = weibullS.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
     mydata->datTime = dataclass.datTime.memptr();
 
     unsigned int l = k;
@@ -530,7 +533,8 @@ void ARMS_Gibbs::arms_gibbs_betaFull(
     mydata->datMu = datMu.memptr();
     mydata->datProportion = datProportion.memptr();
     mydata->weibullS = weibullS.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
     mydata->datTime = dataclass.datTime.memptr();
 
     arma::vec logMu_l = arma::zeros<arma::vec>(N);
@@ -684,7 +688,8 @@ void ARMS_Gibbs::arms_gibbs_zeta(
     mydata->datX = dataclass.datX.memptr();
     mydata->datProportionConst = dataclass.datProportionConst.memptr();
     //mydata->datProportion = datProportion.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
 
     // arma::vec wSq_tmp = arma::ones<arma::vec>(L);
 
@@ -837,7 +842,8 @@ void ARMS_Gibbs::arms_gibbs_zetaK(
     mydata->weibullLambda = weibullLambda.memptr();
     mydata->datX = dataclass.datX.memptr();
     mydata->datProportionConst = dataclass.datProportionConst.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
 
     unsigned int l = k;
     // Gibbs sampling. No need to update variance, since this is only used for updating zetas conditional on proposal etas in M-H sampler
@@ -953,7 +959,8 @@ void ARMS_Gibbs::arms_gibbs_zetaFull(
     mydata->weibullS = weibullS.memptr();
     mydata->weibullLambda = weibullLambda.memptr();
     mydata->datProportionConst = dataclass.datProportionConst.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
 
     // ------------------------------------------------------------------
     // Compute current full alpha matrix once.
@@ -1167,7 +1174,9 @@ void ARMS_Gibbs::arms_kappa(
     mydata->datTheta = datTheta.memptr();
     mydata->datMu = datMu.memptr();
     mydata->datProportion = datProportion.memptr();
-    mydata->datEvent = dataclass.datEvent.memptr();
+    // mydata->datEvent = dataclass.datEvent.memptr();
+    mydata->eventIndex = dataclass.eventIndex.memptr();
+    mydata->nEventIndex = dataclass.eventIndex.n_elem;
     mydata->datTime = dataclass.datTime.memptr();
 
     // define how many ARMS samples to draw for currentPars; first one as intial value