+++ src/drive.cpp
@@ -95,7 +95,9 @@ Rcpp::List run_mcmc(
     unsigned int L = datX.n_slices;
 
     // input constant data sets in a class
-    DataClass dataclass(datEvent, datTime, datX, datX0, datProportionConst);
+    arma::uvec eventIndex = arma::find(datEvent);
+    DataClass dataclass(eventIndex, datTime, datX, datX0, datProportionConst);
+    // DataClass dataclass(datEvent, datTime, datX, datX0, datProportionConst);
 
     // arms parameters in a class
     int metropolis = 1;