+++ src/global.h
@@ -28,7 +28,8 @@ class DataClass
 {
 public:
     // Use const for immutable members
-    const arma::uvec datEvent;
+    // const arma::uvec datEvent;
+    const arma::uvec eventIndex;
     const arma::vec datTime;
     const arma::cube datX;
     const arma::mat datX0;
@@ -43,7 +44,7 @@ public:
 
     // Constructor to initialize the constants
     DataClass(
-        const arma::uvec& datEvent_,
+        const arma::uvec& eventIndex_,
         const arma::vec& datTime_,
         const arma::cube& datX_,
         const arma::mat& datX0_,
@@ -54,7 +55,7 @@ public:
         // std::shared_ptr<arma::mat> weibullLambda_,
         // std::shared_ptr<arma::mat> weibullS_
     ) :
-        datEvent(datEvent_),
+        eventIndex(eventIndex_),
         datTime(datTime_),
         datX(datX_),
         datX0(datX0_),